requirements.txt
click==8.0.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
joblib==1.0.1
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
scikit-learn==0.24.2
scipy==1.7.1
threadpoolctl==2.2.0
